Added test for bin map generation
