+++ tests/conftest.py
@@ -1,0 +1,15 @@
+from parchmint.component import Component
+import pytest
+
+@pytest.fixture
+def component():
+    # Create a component
+    component = Component(
+        name="test_component",
+        ID="test_component",
+        xpos=0,
+        ypos=0,
+        xspan=1000,
+        yspan=1000,
+    )
+    return component
